Améliore la FAQ ainsi que les classes associées
En préparation de l'intégration à la console admin

- Rajoute la propriété de "priorité" à l'entité et la BDD pour les paires questions-réponses
- Mets à jour la BDD
- Renomme \Entities\FAQ ainsi que son fichier (Entities/faq.php) à \Entities\FrequentlyAskedQuestion (Entities/FrequentlyAskedQuestion.php)
- Renomme \Queries\FAQ ainsi que son fichier (Queries/faq.php) à \Queries\FrequentlyAskedQuestions (Queries/FrequentlyAskedQuestions.php)
- Ajoute des commentaires à \Entities\FrequentlyAskedQuestion et \Queries\FrequentlyAskedQuestions
- Répare les définitions de colonnes dans \Queries\FrequentlyAskedQuestions qui bloqueraient la mise à jour des colonnes (manque de gen-on-insert sur les dates)
- Autres changements mineurs
- Nettoie le code
